added permission block when deleting items

+++ backend/src/resolvers/Mutation.js
@@ -43,9 +43,17 @@ const Mutations = {
     async deleteItem(parent, args, ctx, info) {
         const where = { id: args.id };
         // Find the item
-        const item = await ctx.db.query.item({ where }, `{ id title }`);
+        const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
         // Check if they own that item or have permission
-        // TODO
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some
+        (permission => 
+            ['ADMIN', 'ITEMDELETE'].includes(permission)
+        );
+
+        if (!ownsItem && !hasPermission) {
+            throw new Error("You don't have permission to do that!");
+        }
         // Delete it
         return ctx.db.mutation.deleteItem({ where }, info);
     },
